+++ notification/endpoint/http.go
@@ -103,15 +103,13 @@ func (s HTTP) Valid() error {
 			Msg:  "invalid http auth method",
 		}
 	}
-	if s.AuthMethod == "basic" &&
-		(s.Username.Key != s.ID.String()+httpUsernameSuffix ||
-			s.Password.Key != s.ID.String()+httpPasswordSuffix) {
+	if s.AuthMethod == "basic" && (s.Username.Key == "" || s.Password.Key == "") {
 		return &influxdb.Error{
 			Code: influxdb.EInvalid,
 			Msg:  "invalid http username/password for basic auth",
 		}
 	}
-	if s.AuthMethod == "bearer" && s.Token.Key != s.ID.String()+httpTokenSuffix {
+	if s.AuthMethod == "bearer" && s.Token.Key == "" {
 		return &influxdb.Error{
 			Code: influxdb.EInvalid,
 			Msg:  "invalid http token for bearer auth",