+++ ui/src/dashboards/components/DashboardImportOverlay.tsx
@@ -4,6 +4,9 @@ import {withRouter, WithRouterProps} from 'react-router'
 import _ from 'lodash'
 import {connect} from 'react-redux'
 
+// Components
+import ImportOverlay from 'src/shared/components/ImportOverlay'
+
 // Copy
 import {invalidJSON} from 'src/shared/copy/notifications'
 
@@ -15,7 +18,14 @@ import {
 import {notify as notifyAction} from 'src/shared/actions/notifications'
 
 // Types
-import ImportOverlay from 'src/shared/components/ImportOverlay'
+import {ComponentStatus} from '@influxdata/clockface'
+
+// Utils
+import jsonlint from 'jsonlint-mod'
+
+interface State {
+  status: ComponentStatus
+}
 
 interface DispatchProps {
   createDashboardFromTemplate: typeof createDashboardFromTemplateAction
@@ -30,6 +40,10 @@ interface OwnProps extends WithRouterProps {
 type Props = OwnProps & DispatchProps
 
 class DashboardImportOverlay extends PureComponent<Props> {
+  public state: State = {
+    status: ComponentStatus.Default,
+  }
+
   public render() {
     return (
       <ImportOverlay
@@ -37,17 +51,24 @@ class DashboardImportOverlay extends PureComponent<Props> {
         onDismissOverlay={this.onDismiss}
         resourceName="Dashboard"
         onSubmit={this.handleImportDashboard}
+        status={this.state.status}
+        updateStatus={this.updateOverlayStatus}
       />
     )
   }
+
+  private updateOverlayStatus = (status: ComponentStatus) =>
+    this.setState(() => ({status}))
 
   private handleImportDashboard = (uploadContent: string) => {
     const {createDashboardFromTemplate, notify, populateDashboards} = this.props
 
     let template
+    this.updateOverlayStatus(ComponentStatus.Default)
     try {
-      template = JSON.parse(uploadContent)
+      template = jsonlint.parse(uploadContent)
     } catch (error) {
+      this.updateOverlayStatus(ComponentStatus.Error)
       notify(invalidJSON(error.message))
       return
     }